Move previous attribute tracking into trait
